+++ y2021/SonarSweep.cs
@@ -12,7 +12,7 @@ class SonarSweep : IPuzzle
 	internal int CountIncreasesSliding(IEnumerable<int> values, int windowSize = 3)
 	=> CountIncreases(values.Window(windowSize).Select(window => window.Sum()));
 
-	public int Silver(IEnumerable<string> input) => CountIncreases(input.Select(x => Convert.ToInt32(x)));
+	public long Silver(IEnumerable<string> input) => CountIncreases(input.Select(x => Convert.ToInt32(x)));
 	
-	public int Gold(IEnumerable<string> input) => CountIncreases(input.Select(x => Convert.ToInt32(x)).Window(3).Select(window => window.Sum()));
+	public long Gold(IEnumerable<string> input) => CountIncreases(input.Select(x => Convert.ToInt32(x)).Window(3).Select(window => window.Sum()));
 }
